LayoutTests/platform/wpe/imported/w3c/web-platform-tests/resource-timing/same-origin-from-cross-origin-redirect-expected.txt


FAIL Verify that cross origin resources' timings are exposed when the final resource at the end of an HTTP redirect chain is same-origin. assert_equals: redirectStart should be 0 expected 0 but got 15
